+++ internal/rand/eagain.go
