+++ internal/rand/example_test.go
