+++ internal/rand/rand.go
