+++ internal/rand/rand_js.go
