+++ internal/rand/rand_linux.go
