+++ internal/rand/rand_test.go
