+++ internal/rand/rand_windows.go
