+++ internal/rand/util.go
