+++ openpgp/key_generation.go
@@ -8,11 +8,11 @@ import (
 	"crypto"
 	"math/big"
 
+	"github.com/ProtonMail/go-crypto/internal/rsa"
 	"github.com/ProtonMail/go-crypto/openpgp/ecdh"
 	"github.com/ProtonMail/go-crypto/openpgp/errors"
 	"github.com/ProtonMail/go-crypto/openpgp/internal/algorithm"
 	"github.com/ProtonMail/go-crypto/openpgp/packet"
-	"github.com/ProtonMail/go-crypto/rsa"
 	"golang.org/x/crypto/ed25519"
 )
