+++ openpgp/packet/private_key.go
@@ -23,12 +23,12 @@ import (
 	"github.com/ProtonMail/go-crypto/openpgp/internal/ecc"
 	"golang.org/x/crypto/curve25519"
 
+	"github.com/ProtonMail/go-crypto/internal/rsa"
 	"github.com/ProtonMail/go-crypto/openpgp/ecdh"
 	"github.com/ProtonMail/go-crypto/openpgp/elgamal"
 	"github.com/ProtonMail/go-crypto/openpgp/errors"
 	"github.com/ProtonMail/go-crypto/openpgp/internal/encoding"
 	"github.com/ProtonMail/go-crypto/openpgp/s2k"
-	"github.com/ProtonMail/go-crypto/rsa"
 	"golang.org/x/crypto/ed25519"
 )
