+++ openpgp/packet/private_key_test.go
@@ -19,10 +19,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ProtonMail/go-crypto/internal/rsa"
 	"github.com/ProtonMail/go-crypto/openpgp/ecdh"
 	"github.com/ProtonMail/go-crypto/openpgp/elgamal"
 	"github.com/ProtonMail/go-crypto/openpgp/internal/algorithm"
-	"github.com/ProtonMail/go-crypto/rsa"
 	"golang.org/x/crypto/ed25519"
 )
